Update README.txt with new link to download
blitzbasic.com has been down for a long time, and the prebuilt version of Blitz3D is now currently hosted at itch.io. I'm not sure if this repo is still looked at by the owner, it's been a long time since there have been any updates, but it's worth a shot.